refactor: move typewriter speed to ui constants
