--- src/cook-web/src/c-components/ChatUi/chatTypewriterConstants.ts
@@ -1,1 +1,0 @@
-export const CHAT_TYPEWRITER_SPEED_MS = 30;
